libtensorflow-bin: Remove nvidia_x11 dependency
While libcuda.so is definitely a runtime dependency, we cannot assume
that the nixpkgs provided shared library version matches the kernel
module version. We have to deal with this in the same way as we do with
OpenGL drivers and assume the necessary libaries are provided by the
host system (e.g. in /run/opengl-driver/lib on NixOS).