treewide: Don't hardcode specific nvidia_x11
The Nvidia drivers depend on the kernel used. We don't want to hardcode
this dependency but rather use the pattern of linking against
/run/opengl-driver/lib, which makes the packages being independent from
the driver versions.
Because this is also a common pattern with regular CUDA builds the
cudatoolkit ships stubs for all of these shared objects which we make
use of here.