[WIP] cmake/setup-hook.sh: Don't skip build-RPATH
This should simplify using `nix-shell -A` or `nix develop` to develop
CMake based projects. CMake features a mechanism to use a different
RPATH for all executables in the build directory and only rewrite these
RPATHs on installation. This makes it possible to run executables
already from the build directory without having to set LD_LIBRARY_PATH.
This should simplify the checkPhase for cmake based projects and
hopefully not break anything.

Warning: Not tested yet.
TODO:
 - Check if darwin still builds (this was an issue earlier)
 - Check the impact on closure size (this was an issue on an earlier
   try)

Fixes: #22060